refactor: factor out shape generation from `renderElement.ts` pt 2 (#6878)
