+++ src/scene/ShapeCache.ts
@@ -1,28 +1,27 @@
 import { Drawable } from "roughjs/bin/core";
 import { RoughGenerator } from "roughjs/bin/generator";
-import { ExcalidrawElement } from "../element/types";
-import { generateElementShape } from "../renderer/renderElement";
-
-type ElementShape = Drawable | Drawable[] | null;
-
-type ElementShapes = {
-  freedraw: Drawable | null;
-  arrow: Drawable[];
-  line: Drawable[];
-  text: null;
-  image: null;
-};
+import {
+  ExcalidrawElement,
+  ExcalidrawSelectionElement,
+} from "../element/types";
+import { elementWithCanvasCache } from "../renderer/renderElement";
+import { _generateElementShape } from "./Shape";
+import { ElementShape, ElementShapes } from "./types";
 
 export class ShapeCache {
   private static rg = new RoughGenerator();
   private static cache = new WeakMap<ExcalidrawElement, ElementShape>();
 
+  /**
+   * Retrieves shape from cache if available. Use this only if shape
+   * is optional and you have a fallback in case it's not cached.
+   */
   public static get = <T extends ExcalidrawElement>(element: T) => {
     return ShapeCache.cache.get(
       element,
     ) as T["type"] extends keyof ElementShapes
       ? ElementShapes[T["type"]] | undefined
-      : Drawable | null | undefined;
+      : ElementShape | undefined;
   };
 
   public static set = <T extends ExcalidrawElement>(
@@ -41,15 +40,29 @@ export class ShapeCache {
 
   /**
    * Generates & caches shape for element if not already cached, otherwise
-   * return cached shape.
+   * returns cached shape.
    */
-  public static generateElementShape = <T extends ExcalidrawElement>(
+  public static generateElementShape = <
+    T extends Exclude<ExcalidrawElement, ExcalidrawSelectionElement>,
+  >(
     element: T,
+    isExporting = false,
   ) => {
-    const shape = generateElementShape(
+    // when exporting, always regenerated to guarantee the latest shape
+    const cachedShape = isExporting ? undefined : ShapeCache.get(element);
+
+    // `null` indicates no rc shape applicable for this element type,
+    // but it's considered a valid cache value (= do not regenerate)
+    if (cachedShape !== undefined) {
+      return cachedShape;
+    }
+
+    elementWithCanvasCache.delete(element);
+
+    const shape = _generateElementShape(
       element,
       ShapeCache.rg,
-      /* so it prefers cache */ false,
+      isExporting,
     ) as T["type"] extends keyof ElementShapes
       ? ElementShapes[T["type"]]
       : Drawable | null;
